requirements.txt
reportportal-client==3.0.0
robotframework==3.0.2
